+++ src/main/java/pl/lukabrasi/pogodaonline/dtos/WindDto.java
@@ -4,5 +4,5 @@ import lombok.Data;
 
 @Data
 public class WindDto {
-private int speed;
+private double speed;
 }
